refactor(types): name chainable method declarations explicitly
